# app/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dashboard API",
    "version": "1.0.0",
    "description": "API documentation for the Architecture Review Dashboard. \n\n**Authentication**:\n- **Public API**: Requires `x-api-key` header.\n- **Internal API**: Uses Session Cookies."
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session-token"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/ai/architect": {
      "post": {
        "tags": [
          "Public"
        ],
        "summary": "[Public] POST /api/ai/architect",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/repos/{name}": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "[Public] GET /api/repos/{name}",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "tags": [
          "Public"
        ],
        "summary": "[Public] PATCH /api/repos/{name}",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}